Restructured some code and created client connection code
