+++ src/client.h
@@ -1,4 +1,8 @@
 #ifndef CLIENT_H
 #define CLIENT_H
 
+#include "tools.h"
+
+int setup_client(char*, char*);
+
 #endif
