+++ src/tools.h
@@ -1,7 +1,18 @@
 #ifndef TOOLS_H
 #define TOOLS_H
 
+#include <pthread.h>
+
+typedef struct thread{
+	pthread_t id;
+	int socket;
+	int* flag;
+} thread;
+
 void get_primary_ip(char*, size_t);
-int setup_server(char* port);
+
+void* sendMessage(void*);
+void* createMessage(void*);
+void* readMessage(void*);
 
 #endif
